Don't touch any volatile before triggering ptrace
raise() will call memcpy that will destroy some VSX volatiles.
It's possible to trigger a signal without calling raise, for
instance, executing a 'trap' instruction.

After we get into the tracer (parent process), we need to skip
to the next instruction (+4 bytes) in order to avoid getting
the trap in a loop, so we do that using ptrace itself from the
parent process just after getting the first trap.